Move _get_fast_herbie_object() to herbie_utils module
- Rename to get_fast_herbie_object() (remove underscore prefix)
- Function will be used across multiple forecast scripts
- Create herbie_utils.py for reusable Herbie-related utilities
